Move submitVote method
